Initial commit
Took the initial bare-bones script and made it more user-friendly and repeatable.

Given a Spotify user ID and developer credentials (client_id and client_secret) is provided, the user will be able to select any playlist on their account, sort the tracks by the audio_features provided by the Spotify API, and re-arrange the tracks in the playlist all within this script.